+++ src/core/index.ts
@@ -1,6 +1,7 @@
 // Copyright © Aptos Foundation
 // SPDX-License-Identifier: Apache-2.0
 
+export * from "./account";
 export * from "./accountAddress";
 export * from "./authenticationKey";
 export * from "./common";